+++ t/architecture_characterization.t
@@ -15,7 +15,6 @@ sub fresh_device {
     DevIo::reset_test_state();
     %defs = ();
     %attr = ();
-    $modules{Wattpilot}{defptr} = {};
     my $hash = {
         NAME => 'architectureWallbox',
         TYPE => 'Wattpilot',
@@ -25,7 +24,6 @@ sub fresh_device {
         TEST_OPEN => 1,
     };
     $defs{$hash->{NAME}} = $hash;
-    $modules{Wattpilot}{defptr}{$hash->{NAME}} = $hash;
     return $hash;
 }